move the signal handling code out of init into a separate set of files
the goal is to help reduce the amount of random logic that has accumulated in init.c to help make maintenance a bit easier
